changed dim type to int

+++ octagons/tests/libFuzzer/test_oct.h
@@ -14,14 +14,14 @@
 #define R(i) i < K
 
 bool create_octagon(oct_t** octagon, ap_manager_t* man, oct_t * top,
-		long dim, const long *data, size_t dataSize,
+		int dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 bool make_fuzzable(void *array, size_t size, const long *data, size_t dataSize,
 		unsigned int *dataIndex);
 bool assume_fuzzable(bool condition);
 
-bool make_fuzzable_dimension(long * dim, const long *data, size_t dataSize,
+bool make_fuzzable_dimension(int * dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 #endif /* TEST_OCT_H_ */
